📁 feat(refactor): organize stats domain into dedicated package structure
