--- backend/src/main/java/com/dresscode/controller/StatsController.java
@@ -1,25 +1,0 @@
-package com.dresscode.controller;
-
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
-
-import com.dresscode.constants.ApiRoutes;
-import com.dresscode.dto.stats.StatsDto;
-
-import lombok.RequiredArgsConstructor;
-
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.RequestParam;
-
-@RestController
-@RequiredArgsConstructor
-@RequestMapping(ApiRoutes.STATS)
-public class StatsController {
-
-    @GetMapping()
-    public ResponseEntity<StatsDto> getGlobalStats() {
-        return null;
-    }
-
-}
