+++ backend/src/main/java/com/dresscode/service/stats/impl/UserStatsServiceImpl.java
@@ -1,4 +1,4 @@
-package com.dresscode.service.impl;
+package com.dresscode.service.stats.impl;
 
 import java.util.List;
 import java.util.Map;
